feat: replace all cross_lerp with f32lerp, split f32lerp impls
- doc: fix README discord link
- chore: remove Nixon copyright from MIT license
- chore: add --document-private-items to cargo doc workflow
