+++ .github/workflows/rust.yml
@@ -49,7 +49,7 @@ jobs:
       - uses: dtolnay/rust-toolchain@stable
       - uses: Swatinem/rust-cache@v2
       - run: sudo apt-get install --no-install-recommends libasound2-dev libudev-dev
-      - run: cargo doc --workspace --all-features --no-deps
+      - run: cargo doc --workspace --all-features --no-deps --document-private-items
         env:
           RUSTDOCFLAGS: -D warnings
