+++ LICENSE-MIT
@@ -1,7 +1,5 @@
 MIT License
 
-Copyright (c) 2023 Nixon
-
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
 in the Software without restriction, including without limitation the rights
